server/pyproject.toml
[tool.poetry]
name = "norsk-dictionary-server"
version = "0.1.0"
description = ""
authors = ["Thomas Bjerkan <t.bjerkan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1"
responses = "^0.10.7"
beautifulsoup4 = "^4.8"
google-cloud-translate = "^2.0"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
